Add dynamic type definition and update dependencies
Introduced a Define method in Main.cs for dynamic type definition using CSharpCodeProvider and System.Reflection.Metadata. Improved memory recovery validation in MemoryManager.cs. Updated project to C# 9.0, added new package references, and replaced post-build steps with a batch file. Added app.config for assembly binding redirects and updated NuGet dependencies.

+++ Uhara/Code/Interface/MemoryManager.cs
@@ -121,8 +121,12 @@ internal class MemoryManager : MainShared
 
                             if (address == 0 || recover == null) continue;
 
-                            TSaves2.DeleteValue(RegistryName, key, Overwrite, valueName);
                             RefWriteBytes(ProcessInstance, address, recover);
+                            if (!TMemory.ConfirmBytes(ProcessInstance, address, recover))
+                                throw new Exception("Memory recovery exception");
+
+                            TSaves2.DeleteValue(RegistryName, key, Overwrite, valueName);
+
                             //TUtils.Print(recover.Length + " bytes recovered at 0x" + address.ToString("X"));
                         }
                     }
